# Returns an UserAgent struct instead of multiple variables