It makes it easier to use - you don't have to check which return is what, just use correct element of the struct. 

It also makes it easier to add new elements to returned value without breaking backwards compatibility, which I intend to do in later pull requests starting with minor and patch parts of version.