@arussellsaw 2016-10-31
Hey! Thanks for the PR, this is a pretty hefty public api change so i'll wait for @jacobpierce or @ryanslade to come online and give the final approval, but i may as well offer comment on something i've noticed.