@c2h5oh 2016-10-31
Ah yes, I kept that for reference while coding - cleaned up.