@c2h5oh 2016-11-01
@jacobpierce Great! 
This is not the last of it, so if you could provide me with some test data I asked about here https://github.com/avct/uasurfer/issues/18 that would be great. I could use our own logs, but despite large volume I feel the user agent sample would not be diverse enough.