@c2h5oh 2016-11-21
weekly bump

If you don't like any of the proposed changes (either in this pull request or the other two) please voice them, even briefly, so I can work on fixing them (or I know that I have to maintain a fork that fits our needs) - I prefer a rejected pull request to a one hanging in limbo.