@jacobpierce 2016-11-21
@c2h5oh Apologies for taking so long. I really like the direction of this PR, and like seeing structs returned. Looking at performance changes, I've rerun the benchmarks about a dozen times and there wasn't much variation but it did show a 40-60% performance/resource downgrade with the PR:

**Current uasurfer master:**

```
BenchmarkAgentSurfer-16                 200000         10498 ns/op         265 B/op           3 allocs/op
BenchmarkEvalSystem-16                  300000          3760 ns/op           0 B/op           0 allocs/op
BenchmarkEvalBrowserName-16            2000000           604 ns/op           0 B/op           0 allocs/op
BenchmarkEvalBrowserVersion-16          500000          2362 ns/op          15 B/op           0 allocs/op
BenchmarkEvalDevice-16                 2000000           624 ns/op           0 B/op           0 allocs/op
BenchmarkParseChromeMac-16              300000          7163 ns/op         288 B/op           4 allocs/op
BenchmarkParseChromeWin-16              200000          6234 ns/op         240 B/op           3 allocs/op
BenchmarkParseChromeAndroid-16          200000         10624 ns/op         320 B/op           4 allocs/op
BenchmarkParseSafariMac-16              300000          6570 ns/op         288 B/op           4 allocs/op
BenchmarkParseSafariiPad-16             200000          6788 ns/op         288 B/op           4 allocs/op
PASS
ok      code.avct.io/uasurfer    17.671s
```

**PR 'structured' branch:**

```
BenchmarkAgentSurfer-16                 100000         10974 ns/op         330 B/op           5 allocs/op
BenchmarkEvalSystem-16                  300000          3530 ns/op           1 B/op           0 allocs/op
BenchmarkEvalBrowserName-16            3000000           592 ns/op           0 B/op           0 allocs/op
BenchmarkEvalBrowserVersion-16         1000000          2108 ns/op          22 B/op           0 allocs/op
BenchmarkEvalDevice-16                 2000000           586 ns/op           0 B/op           0 allocs/op
BenchmarkParseChromeMac-16              300000          9593 ns/op         384 B/op           6 allocs/op
BenchmarkParseChromeWin-16              200000          8754 ns/op         352 B/op           6 allocs/op
BenchmarkParseChromeAndroid-16          100000         14682 ns/op         432 B/op           6 allocs/op
BenchmarkParseSafariMac-16              200000          9857 ns/op         368 B/op           6 allocs/op
BenchmarkParseSafariiPad-16             200000         11142 ns/op         368 B/op           6 allocs/op
PASS
ok      github.com/c2h5oh/uasurfer    19.402s
```

It would be optimal to return a struct without losing significant performance or raising resource requirements.

I'll add something like this to the readme, but wanted to mention: The existing priorities behind uasurfer are to favor performance, then accuracy and then resolution (e.g. minor version). Currently it's used to evaluate billions of strings per day with low overhead and high turnaround. This is almost opposite to great projects like http://github.com/faisalman/ua-parser-js that can be as slow as they like in favor of accuracy/resolution.

What are your goals for this branch and onward? Ease of use and greater parsing resolution? Perhaps we can work on a structure that can support multiple use cases? Possibly two modes of operation that share tests and logic, one high performance and one high resolution.