@c2h5oh 2016-11-21
Re: goals:
A compromise between the three, favoring speed, for example:

Resolution:
For Android devices major version is insufficient for pretty much anything but showing a metric.
Android doesn't use semantic versioning, so a minor version often means a breaking change. A different API version. Different TLS version supported and a lot of other things. 
Adding minor version support is most likely quite cheap performance wise - one extra cycle in regex state machine, some extra memory pressure, but it is bound to increase memory pressure. Not free, but justifiable considering the wide range of users this is a deal breaker for.

Accuracy: 
99%+ for everything but OS version and 99.9% for 4 out of 6 (with 1 more being close) makes 99.9 or even 99.99 look achievable. I'd be willing to sacrifice 10% of performance for that. Maybe it can be free. If you can share test data I can tinker and see what results I can get.


Re: performance degradation
System, Browser Name/Version and device type are marginally faster or within a margin of error. Os name/version ones seem to have suffered a 50% performance hit - I'll look into that.