@c2h5oh 2016-11-21
I can't replicate the performance degradation you've noticed - based on 10 runs for each version - current master and this pull request

```
# benchcmp -best -changed a27c4e4249e7361e683e87b125f113215f705023.txt 11fce5342a1ac5d653f9c509a4dd167349165d53.txt
benchmark                         old ns/op     new ns/op     delta
BenchmarkAgentSurfer-8            5263          5254          -0.17%
BenchmarkEvalSystem-8             2502          2461          -1.64%
BenchmarkEvalBrowserName-8        543           506           -6.81%
BenchmarkEvalBrowserVersion-8     1180          1185          +0.42%
BenchmarkEvalDevice-8             522           499           -4.41%
BenchmarkParseChromeMac-8         3216          3256          +1.24%
BenchmarkParseChromeWin-8         2956          2928          -0.95%
BenchmarkParseChromeAndroid-8     5887          5862          -0.42%
BenchmarkParseSafariMac-8         3540          3541          +0.03%
BenchmarkParseSafariiPad-8        4258          4289          +0.73%
```

it's most likely because c2h5oh/uasurfer includes more changes that are not part of this pull request