@jacobpierce 2016-11-22
Indeed, didn't run tests on the correct branch, thanks for double checking -- looks good. Really pleased with this PR.